Removed truckloads of whitespace.
The objectAdmin tag now allows the developer to pass in and array of columns (aColumns).
Each array item needs to be a structure with label and webskin keys.
The array will be iterated over and run the defined webskin to generate the HTML for that column. An example of the setup is listed below

<cfset aColumns = arrayNew(1) />

<cfset stColumn = structNew() />
<cfset stColumn.label = "Sample Photo" />
<cfset stColumn.webskin = "displayFirstPhoto" />

<cfset arrayAppend(aColumns, stColumn) />

git-svn-id: https://farcry.jira.com/svn/FC/trunk@1378 8f7d33b6-1012-0410-9a5a-bb7ecce9c363