#FC-677 dmImage will now look for includes in plugins as well as the project. In order for this to function, the path has to be stored in the [include] field.
This can have compatibility issues with old records that simply store the filename assuming that the file is in the project. To combat this issue, we have provided a customtag in core/tags/webskin call include.cfm. Use this tag as a replacement for the <cf_include> tag that will currently be used in your dmInclude webskins.

git-svn-id: https://farcry.jira.com/svn/FC/trunk@1541 8f7d33b6-1012-0410-9a5a-bb7ecce9c363