FC-50 Modified FarcryUD to select from multiple password hashing algorithms
This commit only contains implementations for "none" and "unsalted" (which were already in FarcryUD) however it allows for any number of newer and custom hashing algorithms and overrides.

The default is no hashing with case-insensitive password matching (the normal behaviour for FarCry so far). Unsalted MD5 hashing is included only for backwards compatibility, and provides no real security.

git-svn-id: https://farcry.jira.com/svn/FC/trunk@12077 8f7d33b6-1012-0410-9a5a-bb7ecce9c363