throw an error when an expected webskin is missing
this reverses a previous commit which was an attempt to show a 404 for
pages which did not exist. the problem is that for pages which do exist
but have a missing webskin, a 404 is not the correct response code. if
there are still cases where this particular instance throws a 500 which
should be a 404, then perhaps there is something earlier on in the page
processing lifecycle that needs to be fixed.