(doc) extend object property comparision
Add boolean and list cases to fc object property diff

Function will now highlight differences where the ftType of the field is boolean or list. It determines that "yes", "1" and "true" values are not different as booleans. It also tries to show the list name as the highlighted value for list types.