Now boolean values get converted to a list
Loops through the lFilterFields and if it finds a boolean:
Checks to see if the is a struct in the stFilterMetaData for this type and if not creates the struct.
Checks to see if there are struct keys for ftType,ftList,ftDefault. If no entry exists for these keys it will assign values to show a list dropdown.